Fail test for invalid BST
