+++ src/main/java/com/lperthel/datastructures/scratch.java
@@ -1,4 +1,5 @@
 package com.lperthel.datastructures;
 
 public class scratch {
-	}
+	
+}
