+++ src/main/java/com/lperthel/trees/Solution.java
@@ -1,0 +1,7 @@
+package com.lperthel.trees;
+
+public class Solution {
+    public boolean isValidBST(TreeNode root) {
+    	return true;
+    }
+    }
