+++ src/test/java/com/lperthel/NonLinearDataStructures/BinaryTreeLevelOrderTraversalTest.java
@@ -8,12 +8,15 @@ import java.util.List;
 import org.apache.commons.lang3.StringUtils;
 import org.junit.jupiter.api.Test;
 
+import com.lperthel.trees.BinaryTreeLevelOrderTraversal;
+import com.lperthel.trees.TreeNode;
+
 class BinaryTreeLevelOrderTraversalTest {
 private BinaryTreeLevelOrderTraversal solution = new BinaryTreeLevelOrderTraversal();
 	@Test
 	void testExample1() {
 
-		TreeNode input = TreeNode.generateExample1();
+		TreeNode input = TreeNode.generateNaryTreeExample1();
 				List<Integer> row1 = Arrays.asList(3);
 		List<Integer> row2 = Arrays.asList(9,20);
 		List<Integer> row3 = Arrays.asList(15,7);
